requirements.txt
# Aram Trade Management System — Python Dependencies
# Install: pip install -r requirements.txt
# Python 3.11+ required

# --- Core Trading & Backtesting ---
vectorbt>=0.26.0
backtrader==1.9.78.123

# --- DhanHQ API ---
dhanhq>=2.0.0

# --- Data & Market Data ---
yfinance>=0.2.40
pandas>=2.1.0
numpy>=1.26.0
pyarrow>=14.0
fastparquet>=2024.2.0

# --- Technical Analysis ---
pandas-ta>=0.3.14b0
ta>=0.11.0
scipy>=1.11.0

# --- Machine Learning ---
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.3.0
statsmodels>=0.14.0
PyPortfolioOpt>=1.5.5

# --- Dashboard & Visualization ---
streamlit>=1.35.0
plotly>=5.20.0
altair>=5.2.0
matplotlib>=3.8.0
aiohttp>=3.9.0
websockets>=12.0
httpx>=0.27.0
requests>=2.31.0

# --- Configuration & Environment ---
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
PyYAML>=6.0.1

# --- Logging & Monitoring ---
loguru>=0.7.2
rich>=13.7.0
tqdm>=4.66.0

# --- Database ---
sqlalchemy>=2.0.0

# --- Scheduling ---
schedule>=1.2.1
apscheduler>=3.10.4

# --- Testing ---
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
pytest-mock>=3.12.0
hypothesis>=6.100.0

# --- Code Quality ---
black>=24.0.0
ruff>=0.3.0
mypy>=1.9.0
pre-commit>=3.7.0

# --- Utilities ---
python-dateutil>=2.9.0
pytz>=2024.1
click>=8.1.7
tabulate>=0.9.0
openpyxl>=3.1.0
Jinja2>=3.1.0
